+++ current/dox/convert_xy.c
@@ -1,34 +1,36 @@
 #include "urg_sensor.h"
 #include "urg_utils.h"
+#include <inttypes.h>
 #include <stdio.h>
 #include <math.h>
 
-int main(void)
+int32_t main(void)
 {
-urg_t urg;
-long *length_data = NULL;
-int length_data_size;
-int i;
-// \~japanese ‹——£ƒf[ƒ^‚ğ X-Y À•WŒn‚É•ÏŠ·‚µ‚Ä•\¦‚·‚é
-// \~english Converts data to X-Y coordinates and displays it
+    urg_t urg;
+    int32_t *length_data = NULL;
+    int32_t length_data_size;
+    int32_t i;
+    // \~japanese ï¿½ï¿½ï¿½ï¿½ï¿½fï¿½[ï¿½^ï¿½ï¿½ X-Y ï¿½ï¿½ï¿½Wï¿½nï¿½É•ÏŠï¿½ï¿½ï¿½ï¿½Ä•\ï¿½ï¿½ï¿½ï¿½ï¿½ï¿½
+    // \~english Converts data to X-Y coordinates and displays it
 
-length_data_size = urg_get_distance(&urg, length_data, NULL);
-for (i = 0; i < length_data_size; ++i) {
-    // \~japanese ‚»‚Ì‹——£ƒf[ƒ^‚Ìƒ‰ƒWƒAƒ“Šp“x‚ğ‹‚ßAX, Y ‚ÌÀ•W’l‚ğŒvZ‚·‚é
-    // \~english Gets the angle in radians for range data, and convert to X-Y coordinates
-    double radian;
-    long length;
-    long x;
-    long y;
+    length_data_size = urg_get_distance(&urg, length_data, NULL);
+    for (i = 0; i < length_data_size; ++i)
+    {
+        // \~japanese ï¿½ï¿½ï¿½Ì‹ï¿½ï¿½ï¿½ï¿½fï¿½[ï¿½^ï¿½Ìƒï¿½ï¿½Wï¿½Aï¿½ï¿½ï¿½pï¿½xï¿½ï¿½ï¿½ï¿½ï¿½ßAX, Y ï¿½Ìï¿½ï¿½Wï¿½lï¿½ï¿½ï¿½vï¿½Zï¿½ï¿½ï¿½ï¿½
+        // \~english Gets the angle in radians for range data, and convert to X-Y coordinates
+        double radian;
+        int32_t length;
+        int32_t x;
+        int32_t y;
 
-    radian = urg_index2rad(&urg, i);
-    length = length_data[i];
-    // \todo check length is valid
+        radian = urg_index2rad(&urg, i);
+        length = length_data[i];
+        // \todo check length is valid
 
-    x = (long)(length * cos(radian));
-    y = (long)(length * sin(radian));
-    printf("(%ld, %ld), ", x, y);
-}
-printf("\n");
-return 0;
+        x = (int32_t)(length * cos(radian));
+        y = (int32_t)(length * sin(radian));
+        printf("(%" PRId32 ", %" PRId32 "), ", x, y);
+    }
+    printf("\n");
+    return 0;
 }
